server/requirements.txt
absl-py==1.3.0
alembic==1.8.1
aniso8601==9.0.1
APScheduler==3.9.1.post1
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
backcall==0.2.0
backports.functools-lru-cache==1.6.4
backports.zoneinfo==0.2.1
banal==1.0.6
bcrypt==4.0.1
beautifulsoup4==4.11.1
bidict==0.22.1
cachetools==5.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
comm==0.1.2
contourpy==1.0.6
cycler==0.11.0
dataclasses==0.6
dataset==1.5.2
debugpy==1.5.1
decorator==5.1.1
efficientnet-pytorch==0.7.1
entrypoints==0.4
executing==1.2.0
facenet-pytorch==2.5.2
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
Flask-Migrate==4.0.1
flask-restx==1.0.3
Flask-SQLAlchemy==3.0.2
Flask-SSE==1.0.0
flit_core==3.6.0
fonttools==4.38.0
future==0.18.2
google-auth==2.15.0
google-auth-oauthlib==0.4.6
greenlet==2.0.1
grpcio==1.51.1
idna==3.4
ImageHash==4.3.1
importlib-metadata==5.1.0
importlib-resources==5.10.1
ipykernel==6.19.2
ipython==8.7.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
jupyter_client==7.4.8
jupyter_core==5.1.0
kiwisolver==1.4.4
Mako==1.2.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
nest-asyncio==1.5.6
numpy==1.23.5
oauthlib==3.2.2
packaging==22.0
pandas==1.5.2
parso==0.8.3
pickleshare==0.7.5
Pillow==9.2.0
pip==22.3.1
pkgutil_resolve_name==1.3.10
platformdirs==2.6.0
prompt-toolkit==3.0.36
protobuf==3.9.2
psutil==5.9.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.13.0
PyJWT==2.6.0
pyparsing==3.0.9
pyrsistent==0.19.2
python-dateutil==2.8.2
python-dotenv==0.21.0
python-engineio==4.3.4
python-socketio==5.7.2
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
pywin32==227
PyYAML==6.0
pyzmq==23.2.0
redis==4.4.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.2.0
scipy==1.9.3
seaborn==0.12.1
setuptools==65.5.0
six==1.16.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.45
stack-data==0.6.2
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.1.0
torch==1.7.0
torchvision==0.8.0
tornado==6.2
tqdm==4.64.1
traitlets==5.7.1
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.13
waitress==2.1.2
wcwidth==0.2.5
Werkzeug==2.2.2
wheel==0.37.1
wincertstore==0.2
zipp==3.11.0
